+++ org.alienideology.aibot/src/main/java/listener/BotListener.java
@@ -22,7 +22,7 @@ import net.dv8tion.jda.core.hooks.ListenerAdapter;
  *
  * @author Alien Ideology <alien.ideology at alien.org>
  */
-public class BotListener extends ListenerAdapter implements Runnable {
+public class BotListener implements Runnable {
 
     private Thread t;
     private final String threadName = "Console listener Thread";
@@ -30,20 +30,21 @@ public class BotListener extends ListenerAdapter implements Runnable {
     public BotListener()
     {
         t = new Thread(this,threadName);
-        //t.start();
+    }
+
+    public void startThread() {
+        t.start();
     }
 
     @Override
     public void run() 
     {
         Scanner scanner = new Scanner(System.in);
-        while (true) 
-        {
+        while (true)  {
             String input = scanner.nextLine();
             
             //ShutDown
-            if (input.equals("shutdown")) 
-            {
+            if (input.equals("shutdown")) {
                 try {
                     AIBot.shutdown();
                 } catch (IOException ex) {
@@ -52,21 +53,18 @@ public class BotListener extends ListenerAdapter implements Runnable {
             }
             
             //Test Commands
-            else if(input.startsWith("test"))
-            {
+            else if(input.startsWith("test")) {
                 System.out.println("Test wot?");
             }
             
             //Presence
             //-SetGame
-            else if(input.startsWith("setGame"))
-            {
+            else if(input.startsWith("setGame")) {
                 System.out.println("game set to " + UtilBot.setGame(input.substring(8)));
             }
             
             //-SetStatus
-            else if(input.startsWith("setStatus"))
-            {
+            else if(input.startsWith("setStatus")) {
                 try {
                     OnlineStatus status = UtilBot.setStatus(input.substring(10));
                     System.out.println("Status set to " + status.toString());
@@ -77,15 +75,4 @@ public class BotListener extends ListenerAdapter implements Runnable {
         }
     }
 
-    @Override
-    public void onReady(ReadyEvent e) {
-        System.out.println("Status - Logged in as: " + e.getJDA().getSelfUser().getName());
-    }
-
-    @Override
-    public void onException(ExceptionEvent event) {
-        super.onException(event);
-        if(!event.isLogged())
-            System.out.println(event.getCause());
-    }
 }